Remove dead code: cuda hook backend, gpu.go FFT wrappers, prove cruft, test-only p2 toolkit
