--- backend/accelerated/cuda/doc.go
@@ -1,3 +1,0 @@
-//go:build !cuda
-
-package cuda
