+++ internal/gpu/bls12381/doc.go
@@ -1,5 +1,4 @@
-// Package gpu provides the cgo bindings to libgnark_cuda used by the CUDA
-// accelerated PLONK backend (backend/accelerated/cuda) and the device-resident
-// rho-loop. The implementation is behind the cuda build tag (see gpu.go); this
-// file keeps the package present without the tag.
+// Package gpu provides the cgo bindings to libgnark_cuda used by the
+// device-resident PLONK prover. The implementation is behind the cuda build tag
+// (see gpu.go); this file keeps the package present without the tag.
 package gpu
